fix: fold seed file helper tests to satisfy pkg-file-count
Merge WriteSeedExecutionFile and WriteDynamicExecutionFile tests into
deadcode_contract_test.go so pkg/testutil stays within the 15-file limit.
